+++ src/New/Entity/ShieldClass.h
@@ -17,6 +17,10 @@ public:
 	int ReceiveDamage(args_ReceiveDamage* args);
 	bool CanBeTargeted(WeaponTypeClass* pWeapon);
 
+	void BreakShield(AnimTypeClass* pBreakAnim = nullptr, WeaponTypeClass* pBreakWeapon = nullptr);
+	void SetRespawn(int duration, double amount, int rate, bool resetTimer);
+	void SetSelfHealing(int duration, double amount, int rate, bool resetTimer);
+
 	void KillAnim();
 
 	void AI_Temporal();
@@ -26,9 +30,12 @@ public:
 	void InvalidatePointer(void* ptr);
 
 	double GetHealthRatio();
+	void SetHP(int amount);
 	int GetHP();
 	bool IsActive();
 	bool IsAvailable();
+	bool IsBrokenAndNonRespawning();
+	ShieldTypeClass* GetType();
 
 	static void SyncShieldToAnother(TechnoClass* pFrom, TechnoClass* pTo);
 
@@ -44,12 +51,11 @@ private:
 	void SelfHealing();
 	int GetPercentageAmount(double iStatus);
 
-	void BreakShield();
 	void RespawnShield();
 
 	void CreateAnim();
 
-	void WeaponNullifyAnim();
+	void WeaponNullifyAnim(AnimTypeClass* pHitAnim = nullptr);
 	void ResponseAttack();
 
 	void CloakCheck();
@@ -72,17 +78,26 @@ private:
 	bool Temporal;
 	bool Available;
 
+	double SelfHealing_Warhead;
+	int SelfHealing_Rate_Warhead;
+	double Respawn_Warhead;
+	int Respawn_Rate_Warhead;
+
 	ShieldTypeClass* Type;
 
 	struct Timers
 	{
 		Timers() :
 			SelfHealing{ },
-			Respawn{ }
+			SelfHealing_Warhead { },
+			Respawn{ },
+			Respawn_Warhead { }
 		{ }
 
 		TimerStruct SelfHealing;
+		TimerStruct SelfHealing_Warhead;
 		TimerStruct Respawn;
+		TimerStruct Respawn_Warhead;
 
 	} Timers;
 };